Move Docker support to a separate file. Still needs more work.
