+++ userspace/libsinsp/CMakeLists.txt
@@ -37,6 +37,7 @@ set(SINSP_SOURCES
 	chisel.cpp
 	chisel_api.cpp
 	container.cpp
+	container_docker_common.cpp
 	container_info.cpp
 	ctext.cpp
 	cyclewriter.cpp
